chore: Updated rpc getrawtransaction (#35)
* chore: Refactored functional tests

* chore: Added tx_type field in getrawtransaction

* chore: Corrected withdrawal tx checks

* chore: Refactored verbose call in getrawtransaction tests

* chore: Moved staking_fixtures.py into test_framework/staking_utils.py

* chore: Added tests for NONE tx_type
