+++ src/rpc/rawtransaction.cpp
@@ -28,6 +28,7 @@
 #include <script/sign.h>
 #include <script/signingprovider.h>
 #include <script/standard.h>
+#include <staking/transaction.h>
 #include <uint256.h>
 #include <util/moneystr.h>
 #include <util/strencodings.h>
@@ -46,6 +47,35 @@
  * by the RPCs. This can be overridden with the maxfeerate argument.
  */
 static const CFeeRate DEFAULT_MAX_RAW_TX_FEE_RATE{COIN / 10};
+
+bool checkIfWithdrawalTransaction(const CTransaction& tx)
+{
+    for(const auto& input : CMutableTransaction{tx}.vin) {
+        const CStakesDbEntry& stake = ::ChainstateActive().GetStakesDB().getStakeDbEntry(input.prevout.hash);
+        if(stake.isValid()) {
+            return true;
+        }
+    }
+    return false;
+}
+
+std::string getTransactionType(const CTransaction& tx)
+{
+    if(checkIfWithdrawalTransaction(tx)) {
+        return "STAKING_WITHDRAWAL";
+    }
+    StakingTransactionType txType{CStakingTransactionParser{MakeTransactionRef(tx)}.GetStakingTxType()};
+    switch(txType) {
+        case StakingTransactionType::NONE:
+            return "NONE";
+        case StakingTransactionType::DEPOSIT:
+            return "STAKING_DEPOSIT";
+        case StakingTransactionType::BURN:
+            return "STAKING_BURN";
+        default:
+            return "NONE";
+    }
+}
 
 void TxToJSON(const CTransaction& tx, const uint256 hashBlock, UniValue& entry)
 {
@@ -66,6 +96,7 @@ void TxToJSON(const CTransaction& tx, const uint256 hashBlock, UniValue& entry)
                 entry.pushKV("confirmations", 1 + ::ChainActive().Height() - pindex->nHeight);
                 entry.pushKV("time", pindex->GetBlockTime());
                 entry.pushKV("blocktime", pindex->GetBlockTime());
+                entry.pushKV("tx_type", getTransactionType(tx));
             }
             else
                 entry.pushKV("confirmations", 0);
@@ -151,6 +182,7 @@ static UniValue getrawtransaction(const JSONRPCRequest& request)
                              {RPCResult::Type::NUM, "confirmations", "The confirmations"},
                              {RPCResult::Type::NUM_TIME, "blocktime", "The block time expressed in " + UNIX_EPOCH_TIME},
                              {RPCResult::Type::NUM, "time", "Same as \"blocktime\""},
+                             {RPCResult::Type::STR, "tx_type", "Transaction type, including staking and governance types"},
                         }
                     },
                 },
