+++ test/functional/feature_staking_burn.py
@@ -1,17 +1,13 @@
-import decimal
-
 from test_framework.messages import COIN
+from test_framework.staking_utils import BurnStakingTransactionsMixin
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import (
     connect_nodes,
     disconnect_nodes,
 )
 
-STAKING_TX_HEADER = 0x53
-STAKING_TX_BURN_SUBHEADER = 0x42
 
-
-class StakingBurnTest(BitcoinTestFramework):
+class StakingBurnTest(BitcoinTestFramework, BurnStakingTransactionsMixin):
     def set_test_params(self):
         self.num_nodes = 2
 
@@ -21,41 +17,6 @@ class StakingBurnTest(BitcoinTestFramework):
 
     def get_staking_pool_balance(self, node_num: int) -> int:
         return self.nodes[node_num].getstakinginfo()['staking_pool']
-
-    @staticmethod
-    def create_staking_burn_tx_input(change_address: str, amount_to_burn: decimal.Decimal, utxo: dict):
-        fee = COIN // 1000
-        tx_input = {
-            "txid": utxo["txid"],
-            "vout": utxo["vout"]
-        }
-        # tx outputs
-        amount_to_burn_bin = int(amount_to_burn).to_bytes(8, "little")
-        tx_output_staking_burn_header = {
-            "data": (bytes([STAKING_TX_HEADER, STAKING_TX_BURN_SUBHEADER]) + amount_to_burn_bin).hex()
-        }
-        change_amount = (utxo["amount"] * COIN - amount_to_burn - fee) / COIN
-        if change_amount > 0.0001:
-            tx_output_change = {
-                change_address: change_amount
-            }
-            return [tx_input], [tx_output_staking_burn_header, tx_output_change]
-        else:
-            return [tx_input], [tx_output_staking_burn_header]
-
-    def send_staking_burn_tx(self, change_address: str, amount_to_burn: decimal.Decimal, node_num: int):
-
-        i = 0
-        while True:
-            unspent = self.nodes[node_num].listunspent()[i]
-            if unspent["amount"] * COIN >= amount_to_burn:
-                break
-            i = i + 1
-        tx_inputs, tx_outputs = self.create_staking_burn_tx_input(change_address, amount_to_burn, unspent)
-        # create, sign and send staking burn transaction
-        raw_tx = self.nodes[node_num].createrawtransaction(tx_inputs, tx_outputs)
-        signed_raw_tx = self.nodes[node_num].signrawtransactionwithwallet(raw_tx)
-        self.nodes[0].sendrawtransaction(signed_raw_tx['hex'])
 
     def staking_burn_simple_test(self):
         starting_height = 200
